// src/app/api/models/reply-aggregate.ts
/* tslint:disable */
export interface ReplyAggregate {
  noOfLoves?: number;
  commentId?: number;
  description?: string;
  firstName?: string;
  lastName?: string;
  liked?: boolean;
  dateAndTime?: string;
  profilePicture?: string;
  profilePictureContentType?: string;
  replyId?: number;
  timeElapsed?: string;
  userId?: string;
}